Move BotService to bot/services/bot directory
